color修正
