+++ qiita_app/lib/widgets/setting_item.dart
@@ -1,24 +1,5 @@
 import 'package:flutter/material.dart';
-
-class SectionTitle extends StatelessWidget {
-  final String title;
-
-  const SectionTitle({
-    Key? key,
-    required this.title,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(16.0, 32.0, 16.0, 8.0),
-      child: Text(
-        title,
-        style: const TextStyle(fontSize: 12, color: Colors.grey),
-      ),
-    );
-  }
-}
+import 'package:qiita_app/constants/app_colors.dart';
 
 class SettingItem extends StatelessWidget {
   final String title;
@@ -35,7 +16,7 @@ class SettingItem extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ListTile(
-      tileColor: Colors.white,
+      tileColor: AppColors.white,
       title: Text(title),
       trailing: showArrow ? const Icon(Icons.arrow_forward_ios) : null,
       onTap: onTap,
