+++ qiita_app/lib/widgets/setting_section_title.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+import '../constants/app_colors.dart';
+
+class SectionTitle extends StatelessWidget {
+  final String title;
+
+  const SectionTitle({
+    Key? key,
+    required this.title,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(16.0, 32.0, 16.0, 8.0),
+      child: Text(
+        title,
+        style: const TextStyle(fontSize: 12, color: AppColors.secondary),
+      ),
+    );
+  }
+}
